requirements.txt
requests~=2.31.0
github~=1.2.7
PyGithub~=2.1.1
tqdm~=4.66.1